Full Code
Section 1 of this document provides an annotated listing of the code used for optimized Swin Transformer for EuroSAT Classification with advanced techniques such as Mixup, CosineAnnealingWarmRestarts, GradScaler for mixed precision, and a custom parameter scheduling function. The code is designed to run within a Python environment (e.g., Kaggle notebooks, Google Colab) that supports PyTorch, timm, and scikit-learn. 
Section 2 provides the code to train each of the 5 ViT architecture models, namely DeIT, YOLO, Swin, PVT and MAE. This was used to select the baseline model used in Section 1 above.